app/requirements.txt
huggingface-hub
numpy
safetensors
einops
gradio==5.49.1
soundfile
openai-whisper